create game sockets class to decouple from the socket server
