+++ frontend/src/Socket.tsx
@@ -1,5 +1,5 @@
 import { io, Socket } from 'socket.io-client';
-import { ServerToClientEvents, ClientToServerEvents } from '../../backend/src/Types.ts';
+import { ServerToClientEvents, ClientToServerEvents } from '../../backend/src/sockets/GameSockets.ts';
 
 // "undefined" means the URL will be computed from the `window.location` object
 const URL = process.env.NODE_ENV === 'production' ? undefined : 'http://localhost:3000';
